pyproject.toml
[project]
name = "webeval"
version = "0.1.0"
description = "A web app to collect human evaluations of LLM outputs."
readme = "README.md"
authors = [
    { name = "Matteo Spanio", email = "spanio@dei.unipd.it" }
]
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.13",
    "django-environ>=0.11",
    "user-agents>=2.2",
    "geoip2>=4.8",
    "markdown>=3.6",
    "matplotlib>=3.9",
    "mutagen>=1.47",
    "django-unfold>=0.40,<1.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=5.0",
    "factory-boy>=3.3",
    "selenium>=4.20",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "tests.py"]
addopts = "-ra"
filterwarnings = [
    "ignore::DeprecationWarning:django.*",
]
markers = [
    "selenium: browser-driven end-to-end tests (require Firefox + geckodriver)",
]
